Improve field table handling
